Add files via upload
Responsive Landing Page- Which is created with the help of html, css and javascript. HTML is used to structure the menu, CSS to style it and JAVASCRIPT to add interactivity, such as changing the background or front color of the menu when it scrolled or when a menu item is hovered over.